feat: create add task button in header, add task create popup
