+++ teamTasker/src/pages/LandingPage/LandingPage.tsx
@@ -1,11 +1,11 @@
-import {memo, useEffect} from "react";
+import { memo, useEffect } from "react";
 import styles from "src/App/App.module.scss";
-import {useAppDispatch} from "src/hooks/storeHooks.ts";
-import {useSelector} from "react-redux";
-import {getState} from "src/entities/User";
-import {userAuth} from "src/entities/User/lib/services/userAuth.ts";
+import { useAppDispatch } from "src/hooks/storeHooks.ts";
+import { useSelector } from "react-redux";
+import { getState } from "src/entities/User";
+import { userAuth } from "src/entities/User/lib/services/userAuth.ts";
 
-export const LandingPage = memo(()=> {
+export const LandingPage = memo(() => {
 
     const dispatch = useAppDispatch();
     const state = useSelector(getState);
@@ -15,7 +15,7 @@ export const LandingPage = memo(()=> {
         dispatch(userAuth())
     }, [dispatch]);
 
-    return(
+    return (
         <>
             <div className={styles.tasks}>Hello</div>
             {state.user.data.isAuth && <div>{state.user.data.username}</div>}
