--- teamTasker/src/pages/TaskCreatePage/TaskCreatePage.module.scss
@@ -1,33 +1,0 @@
-.createTasks {
-  background: var(--background-light);
-  border-radius: 15px;
-  padding: 20px;
-  margin-top: 20px;
-  width: 500px;
-  text-align: center;
-  height: fit-content;
-
-  //overflow: ;
-
-  p{
-    margin: 0;
-    text-align: center;
-
-  }
-
-   p, button{
-    margin-top: 10px;
-  }
-
-  input{
-    width: 94%;
-  }
-
-  button{
-    width: 100%;
-  }
-
-  .submitButton{
-
-  }
-}
