+++ teamTasker/src/pages/index.ts
@@ -2,4 +2,3 @@ export { Login } from "./Authorization/Login/Login.tsx"
 export { SingIn } from "./Authorization/SingIn/SingIn.tsx"
 export { LandingPage } from "./LandingPage/LandingPage.tsx"
 export { ProjectCreatePage } from "./ProjectsPage/ProjectCreatePage.tsx"
-export { TaskCreatePage } from "./TaskCreatePage/TaskCreatePage.tsx"
