+++ teamTasker/src/popups/index.ts
@@ -1,0 +1,2 @@
+export { TaskDetailsPopup } from "./TaskDetailsPopup/TaskDetailsPopup.tsx"
+export { TaskCreationPopup } from "./TaskCreationPopup/TaskCreationPopup.tsx"
