+++ teamTasker/src/router/router.tsx
@@ -1,6 +1,6 @@
 import { createBrowserRouter } from "react-router-dom";
 import App from "src/App/App.tsx";
-import { LandingPage, Login, ProjectCreatePage, SingIn, TaskCreatePage } from "src/pages";
+import { LandingPage, Login, ProjectCreatePage, SingIn } from "src/pages";
 import { ProjectsPage } from "src/pages/ProjectsPage/ProjectsPage.tsx";
 import { Project } from "src/entities/Project/ui/Project.tsx";
 import { ProtectedRoute } from "src/shared/ProtectedRoute/ProtectedRoute.tsx";
@@ -36,13 +36,6 @@ export const getRouter = () => {
                         </ProtectedRoute>
                 },
                 {
-                    path: "addTasksToProject/:projectId",
-                    element:
-                        <ProtectedRoute>
-                            <TaskCreatePage/>
-                        </ProtectedRoute>
-                },
-                {
                     path: "workspace",
                     element:
                         <ProtectedRoute>
@@ -55,7 +48,6 @@ export const getRouter = () => {
                         <ProtectedRoute>
                             <Project/>
                         </ProtectedRoute>
-
                 }
             ]
         },
