+++ teamTasker/src/shared/Input/ui/Input.tsx
@@ -1,4 +1,5 @@
 import React, { InputHTMLAttributes } from 'react';
+import styles from "./Input.module.scss";
 
 interface InputProps extends InputHTMLAttributes<HTMLInputElement> {
     setValue?: (value: string) => void; // Accept the input value as an argument
@@ -9,5 +10,5 @@ export const Input: React.FC<InputProps> = ({ ...rest }) => {
     //     setValue(event.target.value); // Pass the input value to the setValue function
     // };
 
-    return <input {...rest} />;
+    return <input className={styles.input} {...rest} />;
 };
